Use DISTRHO_PLUGIN_NAME for VST getLabel()
That value is used as a "product string" by the VST wrapper.
Most other VST plugins seem to use the full plugin name instead.

I have no idea what the difference is, but it doesn't feel right
to use the restricted label used by LADSPA and DSSI for this purpose🤷‍♀️